python/requirements.txt
mss==9.0.1
Pillow==10.3.0
pystray==0.19.5